refactor(atmo): address PR #228 review — move ISA helpers + i18n popup
- Move altitudeToPressure / pressureToAltitudeFt / iasToTasISA out of
  visualization/skewt/atmo-utils.ts into a neutral ts/utils/atmo.ts so the
  dependency direction is top-level → utils (not top-level →
  visualization/skewt). Update the four intra-skewt importers
  (axes, interaction, overlay-bands) plus flights-main and the renamed
  atmo.test.ts. No logic change.
- i18n the cruise-speed (i) popup: route title + body through t() with new
  settings.cruiseSpeedInfoTitle / settings.cruiseSpeedInfoBody keys across
  en/fr/de/es, so non-English users no longer see hard-coded English.
