+++ web/ts/i18n/locales/en.json
@@ -729,6 +729,8 @@
   "page.settings.cruiseAltitude": "Cruise Altitude (ft)",
   "page.settings.flightCeiling": "Flight Ceiling (ft)",
   "page.settings.speed": "Cruise Speed (IAS in kt)",
+  "settings.cruiseSpeedInfoTitle": "Cruise Speed (IAS)",
+  "settings.cruiseSpeedInfoBody": "<p>Enter your cruise speed as <strong>indicated airspeed (IAS)</strong> in knots — the number you read off the ASI at cruise power.</p><p>For time and distance estimates we convert IAS to <strong>true airspeed (TAS)</strong>, which is higher at altitude (roughly +2% per 1,000 ft). We use the <strong>ISA standard atmosphere</strong> at your selected cruise altitude, so changing the altitude updates the estimate.</p>",
   "page.settings.forecastModels": "Forecast Models",
   "page.settings.forecastModelsHint": "Select which NWP models to use for briefings with this profile.",
   "page.settings.optionalServices": "Optional Services",
